Move pkg/kubectl/metricsutil to staging
