+++ staging/publishing/import-restrictions.yaml
@@ -132,6 +132,7 @@
   - k8s.io/client-go
   - k8s.io/kubectl
   - k8s.io/kube-openapi
+  - k8s.io/metrics
   - k8s.io/utils
   - k8s.io/klog
