+++ staging/publishing/rules.yaml
@@ -626,6 +626,10 @@ rules:
       branch: master
     - repository: client-go
       branch: master
+    - repository: code-generator
+      branch: master
+    - repository: metrics
+      branch: master
   - source:
       branch: release-1.15
       dir: staging/src/k8s.io/kubectl
@@ -638,3 +642,7 @@ rules:
       branch: release-1.15
     - repository: client-go
       branch: release-12.0
+    - repository: code-generator
+      branch: release-1.15      
+    - repository: metrics
+      branch: release-1.15
