+++ staging/src/k8s.io/kubectl/go.mod
@@ -31,6 +31,7 @@ require (
 	k8s.io/client-go v0.0.0
 	k8s.io/klog v0.3.1
 	k8s.io/kube-openapi v0.0.0-20190709113604-33be087ad058
+	k8s.io/metrics v0.0.0
 	k8s.io/utils v0.0.0-20190607212802-c55fbcfc754a
 	sigs.k8s.io/yaml v1.1.0
 )
@@ -46,5 +47,7 @@ replace (
 	k8s.io/apimachinery => ../apimachinery
 	k8s.io/cli-runtime => ../cli-runtime
 	k8s.io/client-go => ../client-go
+	k8s.io/code-generator => ../code-generator
 	k8s.io/kubectl => ../kubectl
+	k8s.io/metrics => ../metrics
 )
