+++ staging/src/k8s.io/kubectl/pkg/metricsutil/BUILD
@@ -1,7 +1,4 @@
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
@@ -9,10 +6,9 @@ go_library(
         "metrics_client.go",
         "metrics_printer.go",
     ],
-    importpath = "k8s.io/kubernetes/pkg/kubectl/metricsutil",
-    visibility = [
-        "//build/visible_to:pkg_kubectl_metricsutil_CONSUMERS",
-    ],
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/kubectl/pkg/metricsutil",
+    importpath = "k8s.io/kubectl/pkg/metricsutil",
+    visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/api/resource:go_default_library",
@@ -32,13 +28,12 @@ filegroup(
     name = "package-srcs",
     srcs = glob(["**"]),
     tags = ["automanaged"],
+    visibility = ["//visibility:private"],
 )
 
 filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
-    visibility = [
-        "//build/visible_to:pkg_kubectl_metricsutil_CONSUMERS",
-    ],
+    visibility = ["//visibility:public"],
 )
